requirements_agent.txt
# Farmatodo Agent - Requirements
# Install: pip install -r requirements_agent.txt

# Browser automation
playwright>=1.40.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# AI
google-generativeai>=0.3.0

# Database
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0

# After installing, run:
# playwright install chromium
